use-cases/agent-factory-with-subagents/agents/online_support_content_architect/requirements.txt
# Psychology Content Architect Agent - Dependencies

# === CORE PYDANTIC AI FRAMEWORK ===
pydantic-ai[openai,gemini]>=0.0.13

# === DATA VALIDATION AND SETTINGS ===
pydantic>=2.5.0
pydantic-settings>=2.1.0

# === ENVIRONMENT AND CONFIGURATION ===
python-dotenv>=1.0.0

# === ASYNC AND HTTP SUPPORT ===
httpx>=0.25.0
aiohttp>=3.9.0

# === DATA PROCESSING FOR PROGRAM ARCHITECTURE ===
numpy>=1.24.0
pandas>=2.0.0

# === ARCHITECTURE AND DESIGN PATTERN SUPPORT ===
networkx>=3.1  # For program flow and dependency analysis
graphviz>=0.20.1  # For visualization of program architectures

# === TEXT PROCESSING FOR CONTENT ORGANIZATION ===
nltk>=3.8.0
textstat>=0.7.3

# === DATE AND TIME PROCESSING ===
python-dateutil>=2.8.0

# === JSON PROCESSING ===
orjson>=3.9.0

# === TESTING (OPTIONAL) ===
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# === LOGGING AND MONITORING (OPTIONAL) ===
structlog>=23.1.0
rich>=13.5.0

# === ARCHITECTURE SPECIFIC PACKAGES ===
# Program structure validation
cerberus>=1.3.4  # Schema validation for program structures
jsonschema>=4.19.0  # JSON schema validation

# Template and pattern support
jinja2>=3.1.0  # Template engine for program generation
pyyaml>=6.0  # YAML support for configuration files

# === PSYCHOLOGY AND RESEARCH SPECIFIC (OPTIONAL) ===
# Evidence-based practice support
scipy>=1.11.0  # Statistical analysis for evidence validation
scikit-learn>=1.3.0  # Machine learning for optimization

# === ARCHON MCP INTEGRATION ===
# Note: These are imported dynamically and not strict requirements
# mcp-archon  # Archon MCP Server integration (if available)

# === PROGRAM DESIGN AND ARCHITECTURE TOOLS (OPTIONAL) ===
# Workflow and process design
# pygraphviz>=1.10  # Advanced graph visualization (requires Graphviz system package)
# plantuml>=0.3.0  # UML diagram generation

# === EDUCATIONAL DESIGN PATTERNS (OPTIONAL) ===
# Learning design and instructional patterns
# bloom-taxonomy>=1.0.0  # Bloom's taxonomy implementation
# learning-objectives>=0.1.0  # Learning objectives framework

# === THERAPEUTIC MODALITY SUPPORT (OPTIONAL) ===
# CBT patterns and structures
# cbt-toolkit>=0.1.0  # CBT-specific tools and patterns
# mindfulness-interventions>=0.1.0  # Mindfulness-based intervention patterns

# === CULTURAL ADAPTATION SUPPORT (OPTIONAL) ===
# Internationalization and localization
babel>=2.12.0  # Internationalization support
langdetect>=1.0.9  # Language detection

# === QUALITY ASSURANCE AND VALIDATION (OPTIONAL) ===
# Program quality metrics
# program-quality-metrics>=0.1.0  # Quality assessment tools
# clinical-guidelines-checker>=0.1.0  # Clinical guidelines validation

# === MINIMUM PYTHON VERSION ===
# Requires Python 3.9+

# === NOTES FOR DEVELOPERS ===
# 1. Graphviz system package required for advanced visualizations
# 2. Some psychology-specific packages may require additional dependencies
# 3. Cultural adaptation packages may require language model downloads
# 4. Quality assurance tools may require clinical guideline databases
# 5. Advanced analysis features may benefit from statistical computing environments

# === RECOMMENDED ADDITIONAL TOOLS (NOT PYTHON PACKAGES) ===
# - Draw.io or Lucidchart for program architecture diagrams
# - ATLAS.ti or NVivo for qualitative program analysis
# - Statistical software (R, SPSS) for advanced outcome analysis
# - Clinical guidelines databases (APA, NICE, etc.)
# - Cultural competency frameworks and resources